Fix http library-level methods: upgrade to latest test package
