+++ node_interop/pubspec.yaml
@@ -11,6 +11,7 @@ dependencies:
   js: ^0.6.1
 
 dev_dependencies:
-  test: ^0.12.24
-  build_runner: ^0.7.9
-  build_node_compilers: ^0.1.0
+  test: ^0.12.35
+  build_runner: ^0.8.2+2
+  build_node_compilers: ^0.1.2
+  build_test:
